REM-69 Add comprehensive development guidelines and enhance torrent service tests
- Introduced a detailed development guidelines document for the Transmission Client project, covering core principles, Go and React/TypeScript specific guidelines, testing practices, and documentation standards.
- Added unit tests for torrent service actions, including starting, stopping, and setting speed limits for torrents.
- Implemented tests for adding and removing torrents, validating download paths, and managing torrent files.
- Enhanced tests for retrieving torrents and their associated files, ensuring proper error handling and configuration management.
- Improved tests for saving settings and download paths, including scenarios for success and error cases.
